generalise audio injection
